Module2/requirements.txt
python-dotenv==1.0.1
loguru==0.7.2
tenacity>=8.2.0
langchain>=0.0.267
langchain-community>=0.0.1
chromadb>=0.4.0
huggingface-hub<0.26.0
sentence-transformers>=2.2.2
numpy>=1.24.0
google-generativeai>=0.3.0
aiohttp>=3.8.0
fastapi>=0.103.0
uvicorn>=0.23.0
watchdog>=3.0.0
requests>=2.31.0
python-multipart>=0.0.6
scikit-learn>=1.3.0